UPDATED weather code to be an integer and no longer a float

+++ app/models/Reading.java
@@ -5,13 +5,13 @@ import play.db.jpa.Model;
 
 @Entity
 public class Reading extends Model   {
-    public float code;
+    public int code;
     public float temperature;
     public float windSpeed;
     public float windDirection;
     public float pressure;
 
-    public Reading(float code,float temperature,float windSpeed, float windDirection ,float pressure){
+    public Reading(int code,float temperature,float windSpeed, float windDirection ,float pressure){
         this.code = code;
         this.temperature = temperature;
         this.windSpeed = windSpeed;
